Chore/create escrow (#34)

+++ anchor/programs/depo/src/constants.rs
@@ -1,3 +1,3 @@
 pub const SOL_TO_LAMPORTS: u64 = 1_000_000_000;
 pub const MAX_MODULES: u8 = 10;
-pub const ANCHOR_DISCRIMINATOR: u8 = 8;
+pub const ANCHOR_DISCRIMINATOR: usize = 8;
